+++ services/execsrv/informers.go
@@ -9,123 +9,6 @@ import (
 	"github.com/unweave/unweave/api/types"
 )
 
-type heartbeatInformer struct {
-	execID    string
-	observers map[string]HeartbeatObserver
-	mu        sync.Mutex
-	driver    Driver
-	maxFail   int
-	failCount int
-}
-
-// NewPollingHeartbeatInformerFunc returns a HeartbeatInformerFunc that polls the driver
-// for the exec status. If the driver fails to return the status for maxFail times, the
-// informer will inform all observers that the exec has failed and exit.
-func NewPollingHeartbeatInformerFunc(driver Driver, maxFail int) HeartbeatInformerFunc {
-	// Maintain a map of execs and only create a new informer once per exec
-	execs := map[string]*heartbeatInformer{}
-
-	return func(exec types.Exec) HeartbeatInformer {
-
-		if _, ok := execs[exec.ID]; ok {
-			return execs[exec.ID]
-		}
-
-		inf := &heartbeatInformer{
-			execID:    exec.ID,
-			observers: make(map[string]HeartbeatObserver),
-			mu:        sync.Mutex{},
-			driver:    driver,
-			maxFail:   maxFail,
-			failCount: 0,
-		}
-
-		execs[exec.ID] = inf
-
-		return inf
-	}
-}
-
-func (b *heartbeatInformer) Inform(id string, heartbeat Heartbeat) {
-	b.mu.Lock()
-	defer b.mu.Unlock()
-
-	o, ok := b.observers[id]
-	if !ok {
-		return
-	}
-
-	go o.Update(heartbeat)
-}
-
-func (b *heartbeatInformer) Register(o HeartbeatObserver) {
-	b.mu.Lock()
-	defer b.mu.Unlock()
-
-	if _, ok := b.observers[o.ID()]; ok {
-		return
-	}
-	b.observers[o.ID()] = o
-}
-
-func (b *heartbeatInformer) Unregister(o HeartbeatObserver) {
-	b.mu.Lock()
-	defer b.mu.Unlock()
-	delete(b.observers, o.ID())
-}
-
-func (b *heartbeatInformer) Watch() {
-
-	log.Info().
-		Str(types.ExecIDCtxKey, b.execID).
-		Msgf("Starting heartbeat informer for exec %s", b.execID)
-
-	go func() {
-		defer func() {
-			log.Info().
-				Str(types.ExecIDCtxKey, b.execID).
-				Msgf("Heartbeat informer stopped for exec %s", b.execID)
-		}()
-
-		for {
-			select {
-			case <-time.After(10 * time.Second):
-
-				status, err := b.driver.ExecGetStatus(context.Background(), b.execID)
-				if err != nil {
-					b.failCount++
-
-					if b.failCount > b.maxFail {
-						// No need to notify. Heartbeat just stops.
-						log.Info().
-							Str(types.ExecIDCtxKey, b.execID).
-							Msgf("Heartbeat not detected for exec %q", b.execID)
-
-						return
-					}
-
-					continue
-				}
-
-				b.failCount = 0
-
-				for _, o := range b.observers {
-					b.Inform(o.ID(), Heartbeat{
-						ExecID: b.execID,
-						Time:   time.Now(),
-						Status: status,
-					})
-				}
-
-			case <-time.After(2 * time.Minute):
-				log.Info().
-					Str(types.ExecIDCtxKey, b.execID).
-					Msgf("Heartbeat informer still running for exec %s", b.execID)
-			}
-		}
-	}()
-}
-
 type pollingStateInformer struct {
 	execID     string
 	store      Store
